refactor: stuffCategory enum 클래스 파일위치 변경
