+++ src/modules/board/dto/board-create-request.dto.ts
@@ -1,12 +1,14 @@
-import { IsEnum, IsNumber, IsString } from 'class-validator';
-import { StuffCategory } from '../enums/stuffCategory.enum';
+import { IsEnum, IsNotEmpty, IsNumber, IsString } from 'class-validator';
+import { StuffCategory } from '../../../types/enums/stuffCategory.enum';
 import { Type } from 'class-transformer';
 
 export class CreateBoardRequestDto {
   @IsString()
+  @IsNotEmpty()
   stuffName: string;
 
   @IsString()
+  @IsNotEmpty()
   stuffContent: string;
 
   @IsNumber()
@@ -14,11 +16,12 @@ export class CreateBoardRequestDto {
   stuffPrice: number;
 
   @IsEnum(StuffCategory)
+  @IsNotEmpty()
   stuffCategory: StuffCategory;
 
-  @IsNumber()
-  @Type(() => Number)
-  creatorId: number;
+  @IsNotEmpty()
+  @IsString()
+  detailAdress: string;
 
   image: Express.Multer.File;
 }
