+++ src/modules/board/dto/board-pagination-request.dto.ts
@@ -1,6 +1,6 @@
 import { Type } from "class-transformer";
 import { IsEnum, IsInt, IsOptional, IsString} from "class-validator";
-import { StuffCategory } from "../enums/stuffCategory.enum";
+import { StuffCategory } from "../../../types/enums/stuffCategory.enum";
 
 export class BoardPaginationRequestDto{
