+++ src/modules/board/entity/board.entity.ts
@@ -1,8 +1,8 @@
 import { User } from '../../user/entity/user.entity';
 import { BaseEntity } from '../../../global/common/base.entitiy';
-import { StuffCategory } from '../enums/stuffCategory.enum';
 import { Column, Entity, JoinColumn, ManyToOne, OneToMany, PrimaryGeneratedColumn, VersionColumn } from 'typeorm';
 import { Comment } from '../../comment/entity/comment.entity';
+import { StuffCategory } from '../../../types/enums/stuffCategory.enum';
 
 @Entity()
 export class Board extends BaseEntity {
