+++ src/types/enums/stuffCategory.enum.ts
@@ -1,0 +1,16 @@
+export enum StuffCategory {
+    FASHION = '패션의류/잡화',
+    BEAUTY = '뷰티',
+    BABY = '출산/유아동',
+    FOOD = '식품',
+    KITCHEN = '주방용품',
+    HOUSEHOLD = '생활용품',
+    HOMEAPP = '홈인테리어',
+    SPORT = '가전디지털',
+    CAR = '스포츠/레저',
+    DVD = '자동차용품',
+    HOBBY = '완구/취미',
+    OFFICE = '문구/오피스',
+    ANIMAL = '반려동물용품',
+    HEALTH = '헬스/건강식품',
+}
